Add first div of five with skeleton of other four
The first div is finished, however I need to edit it
in order to make the growing and shrinking of the
different elements to fit my OCD.

I will also customize it how I like. The other four
divs are created in what I think is the proper structure.
CSS is what's needed.